requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Azure Storage
azure-storage-file-datalake>=12.14.0

# HTTP Requests
requests>=2.31.0

# Environment Variables (för att hantera nycklar och konfiguration)
python-dotenv>=1.0.0

# Jupyter Notebook Support (om du kör notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Machine Learning & Visualisering
scikit-learn>=1.5.0
xgboost>=2.1.0
seaborn>=0.13.2
matplotlib>=3.8.0
joblib>=1.3.0

